Organize the structure of components
